fix(android-sdk): SPEC-401-A R59 batch — 10 parity gaps closed across 3 lenses
Lens B P1: routeOutcome continue+edgeTarget now detects end_*/nodeType=end
(OnboardingActivity.kt:631) — was falling through to step lookup, missing,
then re-running the SAME step's rules and re-firing paywall_trigger up to
8x before depth>8 force-completed with wrong analytics. Matches iOS
OnboardingRenderer.swift:1169-1172 navigateToTarget end-detection.

Lens B P3: total_duration_ms emitted as Int matching iOS Int(timeIntervalSince
* 1000) — was raw Long from System.currentTimeMillis() diff causing BQ
schema drift between iOS and Android telemetry.

Lens A P1 #2: social_login provider icons styled per iOS — google "G" 18sp
Bold (was 18sp regular), facebook "f" 20sp Bold + Facebook brand blue
#1877F2 when no monochrome icon_style override (was 18sp regular monochrome,
losing brand identity in light buttons).

Lens A P2 #3: ValuePropStep title↔items 24dp gap inserted as Spacer matching
iOS outer VStack(spacing: 24). Outer Column had no verticalArrangement so
title sat directly against the scroll Column.

Lens A P3 #4: WelcomeStep title+subtitle wrapped in inner Column(spacedBy=12)
matching iOS WelcomeStepView nested VStack(spacing: 12) inside outer
VStack(spacing: 24). Title↔subtitle was 24dp on Android vs 12pt on iOS.

Lens A P3 #5: WheelPickerBlock value formatter Locale.US on both
horizontal + vertical branches — matches iOS String(format:) POSIX locale.
Sister of R27; default JVM locale rendered Arabic-Indic digits ("٢٫٥") on
ar/fa/bn/my locales.

Lens C P2 #1: Slider haptic step-tick across 4 sites — Form Slider field
(FormStepComposable), FormInputSliderBlock single Slider, FormInputRangeSlider
both min+max thumbs (ContentBlockRenderer). Matches iOS native Slider(step:)
auto-emit of UISelectionFeedbackGenerator. Bucket-index gated so continuous
sliders (step=0) stay silent.

Lens C P2 #2: NumericWheelPickerBlock centered-value Text now sets
liveRegion=Polite + contentDescription on both horizontal and vertical
branches — TalkBack announces every wheel selection change matching iOS
native Picker(.wheel) auto-VoiceOver behavior.

Lens C P3 #3: ChatStep skip button color → MaterialTheme.colorScheme.onSurface
.alpha(0.6) — was static Color.Gray which kept same hue across light/dark.
Matches iOS .foregroundColor(.gray) (UIColor.systemGray adaptive).

Lens C P3 #4: QuestionStep CTA disabledContainerColor → accentColor.alpha(0.4)
— was static Color.Gray dropping brand identity in disabled state. Matches
iOS FormStepView .background(Color(hex: ctaBg).opacity(0.4)) pattern.

Imports: added LiveRegionMode + liveRegion to ContentBlockRenderer.

Build verified via Mac bridge: exit_code=0, error_count=0, only pre-existing
warnings.

Streak: 0/3 (R59 returned 12 novel; 10 applied, 2 deferred —
FormFieldBlockStyle.height/font_weight 8-site renderer wiring backlog,
PricingCard shadow y-offset has no first-class Compose API).

+++ src/main/kotlin/ai/appdna/sdk/onboarding/OnboardingFlowManager.kt
@@ -78,7 +78,12 @@ internal class OnboardingFlowManager(
                 ))
             },
             onFlowCompleted = { responses ->
-                val durationMs = System.currentTimeMillis() - startTime
+                // SPEC-401-A R59 (Lens B P3) — emit as Int matching iOS
+                // OnboardingFlowManager.swift:84 `Int(... * 1000)`. Without
+                // .toInt() the event payload carries a Long; downstream BQ
+                // schema + dashboards typed against Int from iOS see type
+                // drift on Android-only telemetry.
+                val durationMs = (System.currentTimeMillis() - startTime).toInt()
                 eventTracker.track("onboarding_flow_completed", mapOf(
                     "flow_id" to flow.id,
                     "total_steps" to flow.steps.size,
